read configuration from yaml
