+++ src/satbias/CMakeLists.txt
@@ -4,5 +4,11 @@
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
 
 ecbuild_add_executable( TARGET  satbias2ioda.x
-                        SOURCES satbias2ioda.cpp
+                        SOURCES SatBiasConverter.cpp GsiSatbiasReader.cpp GsiSatbiasReader.h
                         LIBS    Eigen3::Eigen ${oops_LIBRARIES} ioda_engines )
+
+ecbuild_add_test( TARGET  test_satbias_coding_norms
+                  TYPE    SCRIPT
+                  COMMAND ${CMAKE_BINARY_DIR}/bin/${PROJECT_NAME}_cpplint.py
+                  ARGS    --quiet --recursive ${CMAKE_CURRENT_SOURCE_DIR}
+                  WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/bin )
