+++ src/satbias/CPPLINT.cfg
@@ -1,0 +1,1 @@
+filter=-runtime/int
